Simplify state behavior and management logic.
Change the guaranteed ordering of state calls.
Start is now called as soon as a state is pushed, not when it is first run.
Exit is called as soon as a state is popped or superseded with another.
End is called as soon as state is popped.
This removes the need to store state variables inside a State, and generally simplifies the state
management logic.