Add StateStack.push varargs support.
`push(A,B,C)` is equivalent to `push(C);push(B);push(A)`.

This can make the ordering of states easier to visualize when pushing a group
of states that need to be processed in order (individual calls to push would
have to be provided in reverse order).

For example:
states.push(Awake, Breakfast, Code)
is more clear than
states.push(Code); states.push(Breakfast); states.push(Awake).