Implement SoundBank type.
A SoundBank functions like a SoundEffect, but can play the same sound
overlapping multiple times.

It maintains an internal list of instances, and will create a new instance if
play() is called while all current instances are still playing (up to a
provided limit which defaults to 10).

It is RefCounted and disposes of instances when done.